+++ Sources/XcodeGraph/Models/Target.swift
@@ -56,14 +56,19 @@ public struct Target: Equatable, Hashable, Comparable, Codable, Sendable {
     public var playgrounds: [AbsolutePath]
     public let additionalFiles: [FileElement]
     public var buildRules: [BuildRule]
+    @available(*, deprecated, message: """
+    The prune attribute coupled XcodeGraph to a particular use-case of Tuist and therefore
+    we removed it in favor of using metadata as an in-memory context holder that can be leveraged
+    using conventional tags.
+    """)
     public var prune: Bool
     public let mergedBinaryType: MergedBinaryType
     public let mergeable: Bool
     public let onDemandResourcesTags: OnDemandResourcesTags?
-    public let metadata: TargetMetadata
+    public var metadata: TargetMetadata
     public let type: TargetType
-    /// Package directories
     public let packages: [AbsolutePath]
+    public let buildableFolders: [BuildableFolder]
 
     // MARK: - Init
 
@@ -97,7 +102,8 @@ public struct Target: Equatable, Hashable, Comparable, Codable, Sendable {
         onDemandResourcesTags: OnDemandResourcesTags? = nil,
         metadata: TargetMetadata = .metadata(tags: []),
         type: TargetType = .local,
-        packages: [AbsolutePath] = []
+        packages: [AbsolutePath] = [],
+        buildableFolders: [BuildableFolder] = []
     ) {
         self.name = name
         self.product = product
@@ -129,6 +135,7 @@ public struct Target: Equatable, Hashable, Comparable, Codable, Sendable {
         self.metadata = metadata
         self.type = type
         self.packages = packages
+        self.buildableFolders = buildableFolders
     }
 
     /// Given a target name, it obtains the product name by turning "-" characters into "_" and "/" into "_"
@@ -441,7 +448,8 @@ extension Sequence<Target> {
             prune: Bool = false,
             mergedBinaryType: MergedBinaryType = .disabled,
             mergeable: Bool = false,
-            metadata: TargetMetadata = .test()
+            metadata: TargetMetadata = .test(),
+            buildableFolders: [BuildableFolder] = []
         ) -> Target {
             Target(
                 name: name,
@@ -469,7 +477,8 @@ extension Sequence<Target> {
                 prune: prune,
                 mergedBinaryType: mergedBinaryType,
                 mergeable: mergeable,
-                metadata: metadata
+                metadata: metadata,
+                buildableFolders: buildableFolders
             )
         }
 
@@ -501,7 +510,8 @@ extension Sequence<Target> {
             prune: Bool = false,
             mergedBinaryType: MergedBinaryType = .disabled,
             mergeable: Bool = false,
-            metadata: TargetMetadata = .test()
+            metadata: TargetMetadata = .test(),
+            buildableFolders: [BuildableFolder] = []
         ) -> Target {
             Target(
                 name: name,
@@ -529,7 +539,8 @@ extension Sequence<Target> {
                 prune: prune,
                 mergedBinaryType: mergedBinaryType,
                 mergeable: mergeable,
-                metadata: metadata
+                metadata: metadata,
+                buildableFolders: buildableFolders
             )
         }
 
@@ -554,7 +565,8 @@ extension Sequence<Target> {
             filesGroup: ProjectGroup = .group(name: "Project"),
             dependencies: [TargetDependency] = [],
             rawScriptBuildPhases: [RawScriptBuildPhase] = [],
-            onDemandResourcesTags: OnDemandResourcesTags? = nil
+            onDemandResourcesTags: OnDemandResourcesTags? = nil,
+            buildableFolders: [BuildableFolder] = []
         ) -> Target {
             Target(
                 name: name,
@@ -576,7 +588,8 @@ extension Sequence<Target> {
                 filesGroup: filesGroup,
                 dependencies: dependencies,
                 rawScriptBuildPhases: rawScriptBuildPhases,
-                onDemandResourcesTags: onDemandResourcesTags
+                onDemandResourcesTags: onDemandResourcesTags,
+                buildableFolders: buildableFolders
             )
         }